Incapsulamento GPUContext tramite classi Traits
